fix: GET /api/schedules API ScheduleController.java 로 이전
