+++ backend/src/main/java/com/ice/studyroom/domain/reservation/presentation/ReservationController.java
@@ -3,7 +3,6 @@ package com.ice.studyroom.domain.reservation.presentation;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ice.studyroom.domain.reservation.application.ReservationService;
-import com.ice.studyroom.domain.reservation.domain.entity.Schedule;
 import com.ice.studyroom.domain.reservation.presentation.dto.request.CreateReservationRequest;
 import com.ice.studyroom.domain.reservation.presentation.dto.response.CancelReservationResponse;
 import com.ice.studyroom.domain.reservation.presentation.dto.response.GetMostRecentReservationResponse;
@@ -83,16 +81,6 @@ public class ReservationController {
 		return ResponseEntity
 			.status(StatusCode.OK.getStatus())
 			.body(ResponseDto.of(reservationService.getMyReservationQrCode(resId, authorizationHeader)));
-	}
-
-	@Operation(summary = "스터디룸 일정 조회", description = "스터디룸 예약 가능한 일정을 조회합니다.")
-	@ApiResponse(responseCode = "200", description = "스터디룸 일정 조회 성공")
-	@ApiResponse(responseCode = "500", description = "스터디룸 일정 조회 실패")
-	@GetMapping("/schedules")
-	public ResponseEntity<ResponseDto<List<Schedule>>> getSchedule() {
-		return ResponseEntity
-			.status(HttpStatus.OK)
-			.body(ResponseDto.of(reservationService.getSchedule()));
 	}
 
 	@Operation(summary = "단체 스터디룸 예약", description = "단체 단위로 스터디룸을 예약합니다.")
